add: ad translate field upload and move client component
